[Codegen][NFC] Move encoding structs to IREECodegenTypes.h
It follows the IREE core dialect style, which defines the struct types
in *Types.h.
