+++ compiler/src/iree/compiler/Codegen/Dialect/Codegen/IR/BUILD.bazel
@@ -70,6 +70,7 @@ iree_compiler_cc_library(
         "IREECodegenInterfaces.h.inc",
         "IREECodegenOps.cpp.inc",
         "IREECodegenOps.h.inc",
+        "IREECodegenTypes.h",
         "LoweringConfigEnums.cpp.inc",
         "LoweringConfigEnums.h.inc",
         "UKernelOps.cpp.inc",
